requirements.txt
cycler==0.10.0
Cython==0.23.4
decorator==4.0.6
functional==0.4
matplotlib==2.0.0
numpy==1.13.0
olefile==0.44
Pillow==4.1.1
pkgconfig==1.1.0
protobuf==3.3.0
pycurl==7.43.0
pygobject==3.20.0
pyparsing==2.2.0
python-apt==1.1.0b1
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
scikit-learn==0.18.1
scipy==0.19.0
setuptools==36.0.1
six==1.10.0
sklearn==0.0
tensorflow-gpu==1.1.0
Werkzeug==0.12.2
wheel==0.29.0
